# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API untuk Sistem Manajemen Tugas Akhir dan Skripsi Fakultas Ilmu Komputer Universitas Djuanda.",
        "title": "SIMTAS FILKOM API",
        "termsOfService": "https://filkom.unida.ac.id",
        "contact": {
            "name": "Admin SIMTAS FILKOM",
            "email": "admin@filkom.unida.ac.id"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/archives": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Mencari arsip skripsi (full-text search + filter) — semua user terautentikasi",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Archives"
                ],
                "summary": "Cari arsip",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Full-text search",
                        "name": "q",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Filter tahun lulus",
                        "name": "year",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter bidang keahlian",
                        "name": "field_of_study",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter program studi",
                        "name": "study_program",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter dosen pembimbing",
                        "name": "supervisor_id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Halaman (default 1)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Per halaman (default 20)",
                        "name": "per_page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Daftar arsip",
                        "schema": {
                            "$ref": "#/definitions/response.APIResponse"
                        }
                    }
                }
            }
        },
        "/archives/stats": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Mengambil statistik arsip agregat (Admin + Kaprodi only)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Archives"
                ],
                "summary": "Statistik arsip",
                "responses": {
                    "200": {
                        "description": "Statistik arsip",
                        "schema": {
                            "$ref": "#/definitions/response.APIResponse"
                        }
                    }
                }
            }
        },
        "/archives/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Mengambil detail lengkap arsip (semua user terautentikasi)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Archives"
                ],
                "summary": "Detail arsip",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Archive ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detail arsip",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/usecase.ArchiveDetail"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "Tidak ditemukan",
                        "schema": {
                            "$ref": "#/definitions/response.APIResponse"
                        }
                    }
                }
            }
        },
        "/archives/{id}/download": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Menghasilkan presigned URL download arsip (expired 30 menit, role-based access)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Archives"
                ],
                "summary": "Download arsip",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Archive ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Presigned URL",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/usecase.ArchiveDownloadResult"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "403": {
                        "description": "Mahasiswa non-pemilik ditolak",
                        "schema": {
                            "$ref": "#/definitions/response.APIResponse"
                        }
                    },
                    "404": {
                        "description": "Tidak ditemukan",
                        "schema": {
                            "$ref": "#/definitions/response.APIResponse"
                        }
                    }
                }
            }
        },
...
        "/theses/{thesis_id}/archive": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Shortcut untuk mengambil arsip berdasarkan thesis ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Archives"
                ],
                "summary": "Detail arsip per thesis",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Thesis ID (UUID)",
                        "name": "thesis_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detail arsip",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/usecase.ArchiveDetail"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "Tidak ditemukan",
                        "schema": {
                            "$ref": "#/definitions/response.APIResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Mengarsipkan skripsi final setelah yudisium (Mahasiswa pemilik + Admin)",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Archives"
                ],
                "summary": "Arsipkan skripsi",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Thesis ID (UUID)",
                        "name": "thesis_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "File PDF skripsi final (max 25 MB)",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Abstrak Bahasa Indonesia (min 50 kata)",
                        "name": "abstract_id",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Abstrak Bahasa Inggris",
                        "name": "abstract_en",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Kata kunci dipisah koma (min 3)",
                        "name": "keywords",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Tahun lulus",
                        "name": "graduation_year",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Arsip dibuat",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/usecase.ArchiveDetail"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "409": {
                        "description": "Arsip sudah ada",
                        "schema": {
                            "$ref": "#/definitions/response.APIResponse"
                        }
                    },
                    "422": {
                        "description": "Thesis belum graduated",
                        "schema": {
                            "$ref": "#/definitions/response.APIResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "response.APIResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                },
                "meta": {
                    "$ref": "#/definitions/response.Meta"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "response.Meta": {
            "type": "object",
            "properties": {
                "page": {
                    "type": "integer"
                },
                "per_page": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                },
                "total_pages": {
                    "type": "integer"
                }
            }
        },
...
        "usecase.ArchiveDetail": {
            "type": "object",
            "properties": {
                "abstract_en": {
                    "type": "string"
                },
                "abstract_id": {
                    "type": "string"
                },
                "archived_at": {
                    "type": "string"
                },
                "field_of_study": {
                    "type": "string"
                },
                "file_name": {
                    "type": "string"
                },
                "graduation_year": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "keywords": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "student": {
                    "$ref": "#/definitions/usecase.ThesisStudentBrief"
                },
                "study_program": {
                    "type": "string"
                },
                "supervisors": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/usecase.UserBrief"
                    }
                },
                "thesis_id": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "usecase.ArchiveDownloadResult": {
            "type": "object",
            "properties": {
                "download_url": {
                    "type": "string"
                },
                "expires_in": {
                    "type": "integer"
                }
            }
        },
...
        "usecase.ThesisStudentBrief": {
            "type": "object",
            "properties": {
                "full_name": {
                    "type": "string"
                },
                "nim": {
                    "type": "string"
                }
            }
        },
...
        "usecase.UserBrief": {
            "type": "object",
            "properties": {
                "full_name": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Masukkan token dengan format: Bearer {token}",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}